+++ bfg9000/builtins/regenerate.py
@@ -1,8 +1,25 @@
+from collections import namedtuple
+
+from . import builtin
+from .. import log
+from ..backends import list_backends
 from ..backends.make import writer as make
 from ..backends.ninja import writer as ninja
 from ..build_inputs import build_input
 from ..iterutils import listify
 from ..path import Path
+
+
+def _inputs(build_inputs, env):
+    extra = []
+    if env.mopack:
+        extra = [env.tool('mopack').metadata_file]
+    return build_inputs.bootstrap_paths + listify(env.toolchain.path) + extra
+
+
+def _outputs(build_inputs, env):
+    return ([list_backends()[env.backend].filepath] +
+            [i.path for i in build_inputs['regenerate'].outputs])
 
 
 @build_input('regenerate')
@@ -12,14 +29,41 @@ class Regenerate:
         self.depfile = None
 
 
+# This class is used to help serialize the direct inputs/outputs for the
+# `regenerate` build step so that we can consult it when determining whether to
+# abort regeneration in `find_check_cache`.
+class RegenerateFiles(namedtuple('RegenerateFiles', ['inputs', 'outputs'])):
+    @classmethod
+    def make(cls, build_inputs, env):
+        return RegenerateFiles(_inputs(build_inputs, env),
+                               _outputs(build_inputs, env))
+
+    def to_json(self):
+        return {
+            'inputs': [i.to_json() for i in self.inputs],
+            'outputs': [i.to_json() for i in self.outputs],
+        }
+
+    @classmethod
+    def from_json(cls, data):
+        return cls(
+            [Path.from_json(i) for i in data['inputs']],
+            [Path.from_json(i) for i in data['outputs']],
+        )
+
+
+@builtin.execute_hook()
+def log_regenerating(context):
+    if context.regenerating:
+        log.info('regenerating build files')
+
+
 @make.post_rules_hook
 def make_regenerate_rule(build_inputs, buildfile, env):
     bfg9000 = env.tool('bfg9000')
 
-    extra_deps = []
     if env.mopack:
         mopack = env.tool('mopack')
-        extra_deps = [mopack.metadata_file]
         buildfile.rule(
             target=mopack.metadata_file,
             deps=env.mopack + listify(env.toolchain.path),
@@ -30,9 +74,8 @@ def make_regenerate_rule(build_inputs, buildfile, env):
 
     make.multitarget_rule(
         build_inputs, buildfile,
-        targets=[make.filepath] + build_inputs['regenerate'].outputs,
-        deps=(build_inputs.bootstrap_paths + listify(env.toolchain.path) +
-              extra_deps),
+        targets=_outputs(build_inputs, env),
+        deps=_inputs(build_inputs, env),
         recipe=[bfg9000('regenerate', Path('.'))],
         clean_stamp=False
     )
@@ -45,10 +88,8 @@ def ninja_regenerate_rule(build_inputs, buildfile, env):
     if ninja.features.supported('console', env.backend_version):
         rule_kwargs['pool'] = 'console'
 
-    extra_deps = []
     if env.mopack:
         mopack = env.tool('mopack')
-        extra_deps = [mopack.metadata_file]
         buildfile.rule(
             name='regenerate_deps',
             command=bfg9000('run', initial=True, args=mopack(
@@ -74,8 +115,7 @@ def ninja_regenerate_rule(build_inputs, buildfile, env):
         **rule_kwargs
     )
     buildfile.build(
-        output=[ninja.filepath] + build_inputs['regenerate'].outputs,
+        output=_outputs(build_inputs, env),
         rule='regenerate',
-        implicit=(build_inputs.bootstrap_paths + listify(env.toolchain.path) +
-                  extra_deps)
+        implicit=_inputs(build_inputs, env)
     )