+++ test/unit/builtins/test_file_types.py
@@ -319,7 +319,7 @@ class TestDirectory(TestGenericFile):
             File(srcpath('dir/file.txt')),
             File(srcpath('dir/sub/file2.txt')),
         ])
-        with mock.patch('bfg9000.builtins.find.walk', mock_walk):
+        with mock.patch('bfg9000.path.walk', mock_walk):
             self.assertSameFile(
                 self.context[self.fn](self.filename, include='**/*.txt'),
                 expected
@@ -345,7 +345,7 @@ class TestHeaderDirectory(TestDirectory):
             HeaderFile(srcpath('include/file.hpp'), 'c++'),
             HeaderFile(srcpath('include/sub/file2.hpp'), 'c++'),
         ], langs=['c++'])
-        with mock.patch('bfg9000.builtins.find.walk', mock_walk):
+        with mock.patch('bfg9000.path.walk', mock_walk):
             self.assertSameFile(
                 self.context[self.fn](self.filename, include='**/*.hpp'),
                 expected